refactor(theatron): split 8 files exceeding 800-line limit into sub-modules
Decompose oversized files identified in audit issue #2071:

- view/memory: extract detail renderers to detail.rs (1041→723)
- view/chat: extract streaming renderer to streaming.rs (1021→788)
- view/overlay: extract picker overlays to pickers.rs (955→708)
- update/input: extract tests to tests.rs (956→519)
- update/streaming: extract tests to tests.rs (843→438)
- state/memory: extract data model types to types.rs (805→617)
- core/api/types: extract verification types to verification.rs (882→713)
- bootstrap_tests: split into directory with conditional.rs (949→647)

All sub-modules use pub(super) or pub(crate) re-exports to preserve
existing import paths. No behavioral changes.

Closes #2071

Gate-Passed: kanon 0.1.0
